LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-redirect-preflight.any-expected.txt

CONSOLE MESSAGE: line 29: ReferenceError: Can't find variable: dirname

Harness Error (FAIL), message = ReferenceError: Can't find variable: dirname
